switch to limine and add PIT

+++ src/string.c
@@ -4,7 +4,7 @@ void memcpy(void* dest, void* src, size_t n) {
   char* csrc = (char*)src;
   char* cdest = (char*)dest;
 
-  for (int i = 0; i < n; i++)
+  for (int i = 0; i < (int)n; i++)
     cdest[i] = csrc[i];
 }
